pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "^3.0"
python = "^3.8"
djangorestframework = "^3.13.1"
serializers = "^0.2.4"
pymongo = "^4.2.0"
mongoengine = "^0.24.2"
[tool.poetry.dev-dependencies]
